requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
openai>=1.30.0
httpx>=0.27.0
aiofiles>=23.2.1
python-pptx>=0.6.23
pdfplumber>=0.11.0
python-docx>=1.1.2
Pillow>=10.3.0